Added Data Transfer Object

+++ src/main/java/com/nro/footballmanager/controller/PlayerController.java
@@ -1,38 +1,48 @@
 package com.nro.footballmanager.controller;
 
 import com.nro.footballmanager.entity.Player;
+import com.nro.footballmanager.entity.dto.PlayerDTO;
 import com.nro.footballmanager.service.PlayerServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
-@RequestMapping("/player")
+@RequestMapping("/players")
 public class PlayerController {
 
     @Autowired
     private PlayerServiceImpl playerService;
 
-    @PostMapping("./")
-    public Player savePlayer(@Validated @RequestBody Player player) {
-        return playerService.savePlayer(player);
+    @GetMapping("/")
+    public ResponseEntity<List<Player>> getAllPlayers() {
+        List<Player> players = playerService.findAll();
+        return new ResponseEntity<>(players, HttpStatus.OK);
     }
 
-    @GetMapping("./")
-    public List<Player> fetchPlayersList() {
-        return playerService.fetchPlayersList();
+    @PostMapping("/")
+    public ResponseEntity<Player> savePlayer(@Validated @RequestBody Player player) {
+        return new ResponseEntity<>(playerService.savePlayer(player), HttpStatus.CREATED);
     }
 
-    @PutMapping("./{id}")
-    public Player updatePlayer(@RequestBody Player player, @PathVariable("id") Long playerId){
-        return playerService.updatePlayer(player, playerId);
+    @PutMapping("/{id}")
+    public ResponseEntity<PlayerDTO> updatePlayer(@RequestBody PlayerDTO playerDTO, @PathVariable("id") Long playerId) {
+        if (playerService.getById(playerId).isEmpty()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(PlayerDTO.fromEntity(playerService.updatePlayer(playerId,playerDTO)), HttpStatus.OK);
     }
 
-    @DeleteMapping("./{id}")
-    public String deletePlayer(@PathVariable("id") Long playerId){
-        playerService.deletePlayerById(playerId);
-        return "Player deleted successfully";
+    @DeleteMapping("/{id}")
+    public ResponseEntity<HttpStatus> deletePlayer(@PathVariable("id") Long playerId){
+        if(playerService.getById(playerId).isPresent()){
+            playerService.deletePlayerById(playerId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
